shifted timer into the current action thingy
