+++ client/components/Game.js
@@ -2,14 +2,7 @@ import React from 'react'
 import {drawMap} from '../script/sprites'
 import store, {setName} from '../store'
 import {start, mount} from '../script/game'
-import {
-  ControlPanel,
-  PlayerInfo,
-  CurrentAction,
-  AudioPlayer,
-  TimerBar,
-  EndTurn
-} from './'
+import {ControlPanel, PlayerInfo, CurrentAction, AudioPlayer, EndTurn} from './'
 import socket from '../socket'
 
 class Game extends React.Component {
@@ -47,7 +40,6 @@ class Game extends React.Component {
           <CurrentAction />
           <EndTurn />
           <AudioPlayer />
-          <TimerBar />
         </div>
       </div>
     )
