+++ client/components/index.js
@@ -6,11 +6,8 @@ export {default as Game} from './Game'
 export {default as ControlPanel} from './ControlPanel'
 export {default as PlayerInfo} from './PlayerInfo'
 export {default as CurrentAction} from './CurrentAction'
-
 export {default as AudioPlayer} from './AudioPlayer'
-
-export {default as TimerBar} from './TimerBar'
-
+export {default as Timer} from './Timer'
 // chat components
 export {default as Chat} from './Chat'
 export {default as Message} from './Message'
