+++ package.json
@@ -62,6 +62,7 @@
     "pixi.js": "^5.2.0",
     "prop-types": "^15.7.2",
     "react": "^16.8.6",
+    "react-circular-progressbar": "^2.0.3",
     "react-dom": "^16.8.6",
     "react-loading": "^2.0.3",
     "react-redux": "^7.0.1",
