+++ public/style.css
@@ -578,7 +578,7 @@ h5 {
   position: absolute;
   top: 0.5rem;
   left: 0.5rem;
-  /* display: flex;
+  display: flex;
   flex-direction: wrap;
   align-items: center;
   justify-content: space-between;
@@ -586,7 +586,7 @@ h5 {
   border-radius: 1em;
   border: 2px solid #000;
   min-width: 3em;
-  width: 250px; */
+  width: 250px;
 }
 
 #each-player-info {
@@ -617,16 +617,13 @@ h5 {
   position: absolute;
   top: 0;
   left: 40%;
-  width: 300px;
-  padding: 3px;
-  border: 2px solid black;
-  border-radius: 10px;
+  width: 20%;
+  padding: 5px;
+  border-radius: 0 0 10px 10px;
   background-color: rgba(190, 190, 190, 0.7);
-  font-size: 12px;
   display: flex;
-  flex-direction: column;
   align-items: center;
-  justify-content: center;
+  justify-content: space-evenly;
 }
 
 #currentAction > p {
@@ -638,34 +635,12 @@ h5 {
 }
 
 #timer-bar {
-  position: absolute;
+  /* position: absolute; */
+  width: 6rem;
   bottom: 0;
   right: 0;
 }
 
-#timer-title {
+#PIXIapp {
   display: flex;
-  justify-content: center;
-  background-color: rgba(255, 255, 255, 0.5);
-  border-radius: 10px 10px 0 0;
-}
-
-#timer-title > h2 {
-  font-size: 20px;
-  font-weight: bold;
-}
-
-.progress-bar {
-  height: 12px;
-  width: 300px;
-  border-radius: 50px;
-  border: 1px solid #333;
-  background: rgb(118, 1, 1);
-}
-
-.filler {
-  background: darkgray;
-  height: 100%;
-  border-radius: inherit;
-  transition: width 0.2s ease-in;
 }
